+++ src/Autocomplete/Autocomplete.d.ts
@@ -2,14 +2,16 @@ import * as React from 'react';
 import { FieldProps, FormikProps, FormikValues } from 'formik';
 import { OptionsType } from '..'
 
-export interface FSelectProps {
+export interface AutocompleteProps {
   required?: boolean;
   label?: string;
   field: FieldProps;
   form: FormikProps<FormikValues>;
   options: Array<OptionsType>;
+  fullWidth?: boolean;
+  margin?: 'none' | 'dense' | 'normal';
 }
 
-declare const FSelect: React.ComponentType<FSelectProps>;
+declare const Autocomplete: React.ComponentType<AutocompleteProps>;
 
-export default FSelect;
+export default Autocomplete;